Домашняя работа 5.16(Исправления)
