+++ js/main.js
@@ -1,4 +1,3 @@
-import "./setup.js";
-import "./data.js";
-import "./const.js";
-import "./util.js";
+import { createOffers } from './create-offers.js';
+
+console.log(createOffers());
